ref:Move RecordingSession to its own file.
